split lists
